+++ app/commands/server-launcher.js
@@ -2,10 +2,25 @@
  * @type {{check: module.exports.check}}
  */
 module.exports = {
+
+  /**
+   * @param options
+   */
   serverLauncher: function (options) {
     require('../core/server-process-manager').start(options);
   },
+
+  /**
+   * @param options
+   */
   stopServer: function (options) {
     require('../core/server-process-manager').stopServer(options);
+  },
+
+  /**
+   * @param options
+   */
+  restartServer: function (options) {
+    require('../core/server-process-manager').restartServer(options);
   },
 };