+++ app/core/server-process-manager.js
@@ -1,35 +1,51 @@
 /**
  * server-process-manager.js
  */
-const { exec }          = require('child_process');
-const serverDataService = require('./eqemu-data-service-client.js');
-const pathManager       = require('./path-manager');
-const fs                = require('fs');
-const path              = require('path')
-const os                = require('os')
-const util              = require('util')
-const psList            = require('ps-list');
-const debug             = require('debug')('eqemu-admin:process-manager');
+const { exec, execSync } = require('child_process');
+const serverDataService  = require('./eqemu-data-service-client.js');
+const pathManager        = require('./path-manager');
+const fs                 = require('fs');
+const path               = require('path')
+const os                 = require('os')
+const util               = require('util')
+const psList             = require('ps-list');
+const debug              = require('debug')('eqemu-admin:process-manager');
+const config             = require('./eqemu-config-service')
+
+/**
+ * Constants
+ * @type {number}
+ */
+const MAX_PROCESS_POLL_DELTA_SECONDS_BEFORE_KILL = 60;
+const WATCHDOG_TIMER                             = 1000;
+const LAUNCHER_MAIN_LOOP_TIMER                   = 5 * 1000;
 
 /**
  * @type {{check: module.exports.check}}
  */
 module.exports = {
-  erroredStartsCount : {},
-  hasErroredHaltMessage : {},
-  erroredStartsMaxHalt : 5,
-  minZoneProcesses : 3,
-  processCount : {},
-  serverProcessNames : ['zone', 'world', 'ucs', 'queryserv', 'loginserver'],
-  systemProcessList : {},
-  launchOptions : {},
+  erroredStartsCount: {},
+  hasErroredHaltMessage: {},
+  erroredStartsMaxHalt: 5,
+  minZoneProcesses: 3,
+  processCount: {},
+  serverProcessNames: ['zone', 'world', 'ucs', 'queryserv', 'loginserver'],
+  systemProcessList: {},
+  launchOptions: {},
+  watchDogTimer: null,
+  cancelTimedRestart: false,
+  lastProcessPollTime: Math.floor(new Date() / 1000),
 
   /**
    * Launcher initialization
    * @param options
    */
-  init : function (options) {
+  init: function (options) {
     this.launchOptions = options;
+
+    config.init();
+
+    this.minZoneProcesses = config.getAdminPanelConfig('launcher.minZoneProcesses', 3);
 
     let self = this;
     this.serverProcessNames.forEach(function (process_name) {
@@ -40,12 +56,41 @@ module.exports = {
   },
 
   /**
+   * @returns {number}
+   */
+  getWatchDogPollDelta() {
+    return (Math.floor(new Date() / 1000) - this.lastProcessPollTime)
+  },
+
+  /**
    * @param options
    * @returns {Promise<void>}
    */
-  start : async function (options = []) {
+  start: async function (options = []) {
     this.init(options);
 
+    /**
+     * Shared memory
+     */
+    if (config.getAdminPanelConfig('launcher.runSharedMemory', true)) {
+      execSync('./bin/shared_memory', { cwd: pathManager.emuServerPath }).toString();
+    }
+
+    if (config.getAdminPanelConfig('launcher.runLoginserver', false)) {
+      debug('Running loginserver');
+      this.launchOptions.withLoginserver = true;
+    }
+
+    if (config.getAdminPanelConfig('launcher.runQueryServ', false)) {
+      debug('Running queryserv');
+      this.launchOptions.withQueryserv = true;
+    }
+
+    this.startWatchDog();
+
+    /**
+     * Main launcher loop
+     */
     while (1) {
       await this.pollProcessList();
       await this.getBootedZoneCount();
@@ -54,11 +99,12 @@ module.exports = {
        * @type {exports}
        */
       let self = this;
-      ['loginserver', 'ucs', 'world', 'queryserv'].forEach(function (process_name) {
-        if (self.doesProcessNeedToBoot(process_name)) {
+      for (const process_name of ['loginserver', 'ucs', 'world', 'queryserv']) {
+        if (self.doesProcessNeedToBoot(process_name) && !await self.isProcessRunning(process_name)) {
+          debug('starting unique process [' + process_name + ']')
           self.startProcess(process_name);
         }
-      });
+      }
 
       /**
        * Zone
@@ -67,13 +113,9 @@ module.exports = {
         await self.startProcess('zone');
       }
 
-      console.log('Zone Processes: \'%s\'', this.processCount['zone']);
-      console.log('World Process: \'%s\'', this.processCount['world']);
-      console.log('Loginserver Process: \'%s\'', this.processCount['loginserver']);
-      console.log('UCS Process: \'%s\'', this.processCount['ucs']);
-      console.log('Booted Zone Processes: \'%s\'', this.zoneBootedProcessCount);
+      await this.sleep(LAUNCHER_MAIN_LOOP_TIMER);
 
-      await this.sleep(5000);
+      this.lastProcessPollTime = Math.floor(new Date() / 1000);
     }
   },
 
@@ -81,10 +123,10 @@ module.exports = {
    * @param process_name
    * @returns {boolean}
    */
-  doesProcessNeedToBoot : function (process_name) {
+  doesProcessNeedToBoot: function (process_name) {
     if (this.erroredStartsCount[process_name] >= this.erroredStartsMaxHalt) {
       if (!this.hasErroredHaltMessage[process_name]) {
-        console.error('Process \'%s\' has tried to boot too many (%s) times... Halting attempts', process_name, this.erroredStartsMaxHalt)
+        console.error('Process [%s] has tried to boot too many (%s) times... Halting attempts', process_name, this.erroredStartsMaxHalt)
         this.hasErroredHaltMessage[process_name] = 1;
       }
 
@@ -96,10 +138,16 @@ module.exports = {
 
       return true;
     }
+
     if (process_name === 'loginserver') {
       return this.launchOptions && this.launchOptions.withLoginserver && this.processCount[process_name] === 0;
-
     }
+
+    if (process_name === 'queryserv') {
+      return this.launchOptions && this.launchOptions.withQueryserv && this.processCount[process_name] === 0;
+    }
+
+    debug('[doesProcessNeedToBoot] returning [%s]', this.processCount[process_name] === 0)
 
     return this.processCount[process_name] === 0;
   },
@@ -108,20 +156,38 @@ module.exports = {
    * @param zone_short_name
    * @returns {Promise<void>}
    */
-  startStaticZone : async function (zone_short_name) {
+  startStaticZone: async function (zone_short_name) {
     return await this.startProcess('zone', [zone_short_name]);
   },
 
   /**
    * @returns {Promise<void>}
    */
-  stopServer : async function () {
+  stopServer: async function () {
     this.systemProcessList = await psList();
     let self               = this;
 
+    debug('[stopServer] Stopping server...');
+
+    /**
+     * Kill launcher
+     */
     this.systemProcessList.forEach(function (process) {
-      if (self.serverProcessNames.includes(process.name) || process.cmd.includes('server-launcher')) {
+      if (process.cmd.includes('server-launcher')) {
         self.killProcess(process.pid);
+
+        debug('[stopServer] Killing launcher [%s] pid [%s]', process.name, process.pid);
+      }
+    });
+
+    /**
+     * Kill server processes
+     */
+    this.systemProcessList.forEach(function (process) {
+      if (self.serverProcessNames.includes(process.name)) {
+        self.killProcess(process.pid);
+
+        debug('[stopServer] Killing server process [%s] pid [%s]', process.name, process.pid);
       }
     });
 
@@ -129,10 +195,30 @@ module.exports = {
   },
 
   /**
+   * @returns {Promise<boolean>}
+   */
+  async isLauncherBooted() {
+    let isLauncherBooted = false;
+    await this.pollProcessList();
+    this.systemProcessList.forEach(function (process) {
+      if (process.cmd.includes('server-launcher')) {
+        isLauncherBooted = true;
+      }
+    });
+
+    return isLauncherBooted;
+  },
+
+  /**
    * @returns {exports}
    */
-  startServerLauncher : async function (options) {
+  startServerLauncher: async function (options) {
 
+    /**
+     * Parse args
+     *
+     * @type {*[]}
+     */
     let args = [];
     if (options) {
       args.push(options);
@@ -144,19 +230,10 @@ module.exports = {
     });
 
     /**
-     * Check if launcher is booted first
-     * @type {boolean}
+     * Bail out if launcher already started
      */
-    let isLauncherBooted = false;
-    await this.pollProcessList();
-    this.systemProcessList.forEach(function (process) {
-      if (process.cmd.includes('server-launcher')) {
-        isLauncherBooted = true;
-      }
-    });
-
-    if (isLauncherBooted) {
-      console.log('Launcher is already booted');
+    if (await this.isLauncherBooted()) {
+      debug('Launcher is already booted... exiting...');
       return false;
     }
 
@@ -165,24 +242,29 @@ module.exports = {
     // TODO: Windows
     if (process.platform === 'linux') {
       startProcessString = util.format(
-        'PKG_EXECPATH=; %s server-launcher %s &',
+        'while : ; do PKG_EXECPATH=; nohup %s server-launcher %s && sleep 1 ; done &',
+        //'PKG_EXECPATH=; nohup %s server-launcher %s &',
         pathManager.getEqemuAdminEntrypoint(),
         argString
       );
     }
 
-    const command = util.format('rm -rf %s*', this.getLogRedirectDir());
-    exec(command, { cwd : pathManager.emuServerPath });
+    this.purgeServerLogs();
 
-    debug('start string [%s]', startProcessString);
-    debug('cwd [%s]', path.join(path.dirname(process.argv[0]), '../'));
+    debug('[startServerLauncher] Start string [%s]', startProcessString);
 
-    exec(startProcessString,
-      {
-        encoding : 'utf8',
-        cwd : path.join(path.dirname(process.argv[0]), '../')
+    /**
+     * Start launcher
+     */
+    exec(startProcessString, { cwd: pathManager.emuServerPath });
+
+    /**
+     exec(startProcessString,
+     {
+        encoding: 'utf8',
+        // cwd: path.join(path.dirname(process.argv[0]), '../')
       },
-      (error, stdout, stderr) => {
+     (error, stdout, stderr) => {
         if (error) {
           debug(`exec error: ${error}`);
           return;
@@ -190,15 +272,71 @@ module.exports = {
         debug(`stdout: ${stdout}`);
         debug(`stderr: ${stderr}`);
       }
-    );
+     );
+     **/
 
     return this;
   },
 
   /**
+   * @param options
+   */
+  async cancelRestart(options = []) {
+    if (options.cancel) {
+      this.cancelTimedRestart = true;
+      await serverDataService.messageWorld("[SERVER MESSAGE] Server restart cancelled")
+    }
+  },
+
+  /**
    * @returns {exports}
    */
-  restartServer : async function () {
+  restartServer: async function (options = []) {
+
+    /**
+     * Delayed restart
+     */
+    if (options.timer && options.timer > 0) {
+
+      const startTime     = Math.floor(new Date() / 1000);
+      const endTime       = startTime + parseInt(options.timer);
+      let rebootTime      = false;
+      let lastWarningTime = Math.floor(new Date() / 1000) - 1000;
+
+      while (!rebootTime) {
+        const minutesRemaining = (endTime - Math.floor(new Date() / 1000)) / 60;
+        const unixTimeNow      = Math.floor(new Date() / 1000);
+
+        if ((lastWarningTime + 30) <= unixTimeNow) {
+          lastWarningTime = unixTimeNow;
+
+          const worldMessage = util.format(
+            '[SERVER MESSAGE] The world will be coming down for a reboot in [%s] minute(s), please log out before this time...',
+            Number((minutesRemaining).toFixed(2))
+          );
+
+          await serverDataService.messageWorld(worldMessage)
+
+          debug('Sending world message | %s', worldMessage)
+        }
+
+        if (unixTimeNow > endTime) {
+          rebootTime = true;
+        }
+
+        if (this.cancelTimedRestart) {
+          break;
+        }
+
+        await this.sleep(1000);
+      }
+    }
+
+    if (this.cancelTimedRestart) {
+      debug('Reboot cancelled');
+      return false;
+    }
+
     this.stopServer();
 
     let self = this;
@@ -213,7 +351,7 @@ module.exports = {
    * @param ms
    * @returns {Promise<any>}
    */
-  sleep : function (ms) {
+  sleep: function (ms) {
     return new Promise(resolve => {
       setTimeout(resolve, ms)
     });
@@ -222,7 +360,7 @@ module.exports = {
   /**
    * @returns {Promise<number>}
    */
-  getBootedZoneCount : async function () {
+  getBootedZoneCount: async function () {
     const zone_list             = await serverDataService.getZoneList();
     this.zoneBootedProcessCount = 0;
     let self                    = this;
@@ -240,89 +378,47 @@ module.exports = {
     return this.zoneBootedProcessCount;
   },
 
-  getLogRedirectDir() {
-    const logRedirectDir = path.join(os.tmpdir(), 'admin-process-logs');
-    if (!fs.existsSync(logRedirectDir)) {
-      fs.mkdirSync(logRedirectDir);
-    }
-
-    return logRedirectDir;
-  },
-
   /**
    * @param process_name
    * @param args
    * @returns {Promise<void>}
    */
-  startProcess : async function (process_name, args = []) {
+  startProcess: async function (process_name, args = []) {
     let argString = '';
     args.forEach(function (arg) {
       argString += arg + ' ';
     });
 
-    let is_windows           = process.platform === 'win32';
-    let is_linux             = process.platform === 'linux';
-    let start_process_string = '';
-    if (is_windows) {
-      start_process_string = process_name + '.exe';
+    let isWindows          = process.platform === 'win32';
+    let isLinux            = process.platform === 'linux';
+    let startProcessString = '';
+    if (isWindows) {
+      startProcessString = process_name + '.exe';
     }
-    if (is_linux) {
-      start_process_string =
-        util.format('./bin/%s %s 1> %s_$$.txt &',
-          process_name,
-          argString,
-          path.join(this.getLogRedirectDir(), process_name)
-        );
 
-      if (process_name !== 'zone') {
-        const command = util.format('rm -rf %s*', path.join(this.getLogRedirectDir(), process_name));
-        exec(command, { cwd : pathManager.emuServerPath });
-      }
+    if (isLinux) {
+      startProcessString =
+        util.format('./bin/%s %s &',
+          process_name,
+          argString
+        );
     }
 
     debug('Starting process [%s] command [%s] path [%s]',
       process_name,
-      start_process_string,
+      startProcessString,
       pathManager.getEmuServerPath()
     );
 
-    const child = exec(
-      start_process_string,
-      {
-        cwd : pathManager.emuServerPath,
-        encoding : 'utf8'
-      }
-    );
+    exec(startProcessString, { cwd: pathManager.emuServerPath, encoding: 'utf8' });
 
     this.processCount[process_name]++;
-
-    // console.log('stdout ', child);
-  },
-
-  /**
-   * @param process_name
-   * @param args
-   * @param data
-   */
-  handleProcessError : function (process_name, args, data) {
-    console.error('[Error] Process \'%s\' via path: \'%s\', with args failed to start [%s]\n%s',
-      process_name,
-      pathManager.emuServerPath,
-      args,
-      data
-    );
-
-    if (typeof this.erroredStartsCount[process_name] === 'undefined') {
-      this.erroredStartsCount[process_name] = 0;
-    }
-
-    this.erroredStartsCount[process_name]++;
   },
 
   /**
    * @param process_id
    */
-  killProcess : function (process_id) {
+  killProcess: function (process_id) {
     try {
       require('child_process').execSync('kill -9 ' + process_id);
     } catch (e) {
@@ -333,7 +429,7 @@ module.exports = {
   /**
    * @returns {Promise<void>}
    */
-  pollProcessList : async function () {
+  pollProcessList: async function () {
     this.systemProcessList = await psList();
 
     let self = this;
@@ -351,17 +447,91 @@ module.exports = {
   /**
    * @returns {Promise<{world: number, ucs: number, zone: number, queryserv: number, loginserver: number}>}
    */
-  getProcessCounts : async function () {
+  getProcessCounts: async function () {
     await this.pollProcessList();
 
     return {
-      'zone' : this.processCount['zone'],
-      'world' : this.processCount['world'],
-      'ucs' : this.processCount['ucs'],
-      'queryserv' : this.processCount['queryserv'],
-      'loginserver' : this.processCount['loginserver']
+      'zone': this.processCount['zone'],
+      'world': this.processCount['world'],
+      'ucs': this.processCount['ucs'],
+      'queryserv': this.processCount['queryserv'],
+      'loginserver': this.processCount['loginserver']
     };
-  }
+  },
 
-}
-;
+  /**
+   * @param processName
+   * @returns {Promise<void>}
+   */
+  isProcessRunning: async function (processName) {
+    const cmd = (() => {
+      switch (process.platform) {
+        case 'win32' :
+          return `tasklist`;
+        case 'darwin' :
+          return `ps -ax | grep ${processName}`;
+        case 'linux' :
+          return `ps axco command`;
+        default:
+          return false;
+      }
+    })();
+
+    const stdout = require('child_process').execSync(cmd).toString();
+
+    let foundProcess = false;
+    stdout.split('\n').forEach(function (row) {
+      if (row.toLowerCase().trim() === processName.toLowerCase().trim()) {
+        foundProcess = true;
+      }
+    });
+
+    debug('[isProcessRunning] found process [%s] [%s]', processName, foundProcess)
+
+    return foundProcess;
+  },
+
+  /**
+   * Purges server logs
+   */
+  purgeServerLogs() {
+    exec('rm -rf logs/zone/*.log', { cwd: pathManager.emuServerPath });
+    exec('rm -rf logs/*.log', { cwd: pathManager.emuServerPath });
+  },
+
+  /**
+   * Monitors server launcher for communication drift with world to be restarted
+   */
+  startWatchDog() {
+    let self           = this;
+    this.watchDogTimer = setInterval(function () {
+
+      let processStatus = {
+        'Zone Processes': self.processCount['zone'],
+        'Booted Zone Processes': self.zoneBootedProcessCount,
+        'World Processes': self.processCount['world'],
+        'Loginserver Processes': self.processCount['loginserver'],
+        'UCS Processes': self.processCount['ucs'],
+        'Watchdog Polling Delta': self.getWatchDogPollDelta() + ' / ' + MAX_PROCESS_POLL_DELTA_SECONDS_BEFORE_KILL
+      };
+
+      console.table(processStatus);
+
+      if (self.getWatchDogPollDelta() > MAX_PROCESS_POLL_DELTA_SECONDS_BEFORE_KILL) {
+        console.log('HIT WATCHDOG THRESHOLD. TERMINATING...');
+
+        const data    = new Date() + ' watchdog killed launcher\n';
+        const logFile = pathManager.getEmuServerPath() + '/logs/launcher-watchdog-kill.log';
+
+        if (fs.existsSync(logFile)) {
+          fs.appendFileSync(logFile, data);
+        } else {
+          fs.writeFileSync(logFile, data);
+        }
+
+        process.exit()
+      }
+
+    }, WATCHDOG_TIMER);
+  }
+};